move SidebarElement to its own file
